tests(api) convert router tests for koko [KAG-3286] (#7525)
* tests(api) convert router tests for koko

* tests(api) add router test to koko executor

---------

+++ spec-ee/kong-api-tests/test/gateway/admin-api/router.spec.ts
@@ -21,6 +21,8 @@ import {
   getKongContainerName,
   waitForConfigRebuild,
   deleteGatewayRoute,
+  isGateway,
+  isKoko,
 } from '@support';
 
 const agent = new https.Agent({
@@ -46,28 +48,23 @@ const currentHost = getGatewayHost();
 const isLocalDb = isLocalDatabase();
 
 describe('@smoke: Router Functionality Tests', function () {
-  const routesUrl = `${getBasePath({
-    environment: Environment.gateway.adminSec,
-  })}/routes`;
-
-  const adminUrl = `${getBasePath({
-    environment: Environment.gateway.adminSec,
-  })}`;
-
   const proxyUrl = `${getBasePath({
+    app: 'gateway',
     environment: Environment.gateway.proxySec,
   })}`;
 
   const isHybrid = isGwHybrid();
   const serviceName = randomString();
   const kongContainerName = getKongContainerName();
-  const regexPath = '~/(hell?o|world)-(?<user>\\S+)';
+  const regexPath = '~/(hell?o|world)-(?<user>[a-zA-Z]+)';
   const waitTime = 5000;
   const longWaitTime = 10000;
   const hybridWaitTime = 6000;
 
   let serviceDetails: any;
   let routeId: string;
+  let adminUrl: any;
+  let routesUrl: any
 
   const routePayload = {
     name: randomString(),
@@ -76,9 +73,14 @@ describe('@smoke: Router Functionality Tests', function () {
   };
 
   before(async function () {
-    const serviceReq = await createGatewayService(serviceName, {
-      url: 'http://httpbin/anything',
-    });
+    adminUrl = `${getBasePath({
+      environment: isGateway() ? Environment.gateway.admin : undefined,
+    })}`;
+  
+    routesUrl = `${adminUrl}/routes`;
+
+    const serviceReq = await createGatewayService(serviceName);
+
     serviceDetails = {
       id: serviceReq.id,
       name: serviceReq.name,
@@ -127,13 +129,24 @@ describe('@smoke: Router Functionality Tests', function () {
     );
     logResponse(resp);
 
-    expect(resp.status, 'Status should be 409').equal(409);
-    expect(resp.data.name, 'Should have correct error name').equal(
-      'unique constraint violation'
-    );
-    expect(resp.data.message, 'Should have correct error name').equal(
-      `UNIQUE violation detected on '{name="${routePayload.name}"}'`
-    );
+    // *** RESPONSE DIFFERENCES IN GATEWAY AND KOKO ***
+    if (isGateway()) {
+      expect(resp.status, 'Status should be 409').equal(409);
+      expect(resp.data.name, 'Should have correct error name').equal(
+        'unique constraint violation'
+      );
+      expect(resp.data.message, 'Should have correct error name').equal(
+        `UNIQUE violation detected on '{name="${routePayload.name}"}'`
+      );
+    } else if (isKoko()) {
+      expect(resp.status, 'Status should be 400').to.equal(400);
+      expect(resp.data.message, 'Should have correct error name').to.equal(
+        'data constraint error'
+      );
+
+      // in Konnect we need to wait for the route creation to take effect
+      await waitForConfigRebuild()
+    }
   });
 
   testWrongHeaders.forEach(({ header, value }) => {
@@ -233,21 +246,44 @@ describe('@smoke: Router Functionality Tests', function () {
   });
 
   it('should PATCH the route headers and paths', async function () {
+    // *** KOKO DOES NOT PATCH REQUESTS AND HEADERS SHOULD BE NULL ***
     const resp = await axios({
-      method: 'patch',
+      method: isGateway() ? 'patch' : 'put',
       url: `${routesUrl}/${routeId}`,
       data: {
-        headers: [],
-        paths: [regexPath],
+        service: {
+          id: serviceDetails.id
+        },
+        headers: isGateway() ? [] : null,
+        paths: [regexPath]
       },
       headers: { 'Content-Type': 'application/json' },
     });
     logResponse(resp);
 
     expect(resp.status, 'Status should be 200').equal(200);
-    expect(resp.data.headers, 'Should have empty headers').to.be.empty;
     expect(resp.data.paths[0], 'Should have correct path').to.equal(regexPath);
+
+    // *** KOKO DOES NOT CONTAIN HEADERS in RESPONSE ***
+    if(isGateway()) {
+      expect(resp.data.headers, 'Should have empty headers').to.be.empty
+    } else if (isKoko()) {
+      expect(resp.data.headers, 'Should have no headers').to.not.exist
+    }
+
     await wait(isLocalDb ? waitTime : longWaitTime); // eslint-disable-line no-restricted-syntax
+  });
+  
+  regexCorrectPatterns.forEach((correctPattern) => {
+    it(`should route a request with matching regex path: ${correctPattern}`, async function () {
+      const req = () => axios(`${proxyUrl}${correctPattern}`);
+
+      const assertions = (resp) => {
+        expect(resp.status, 'Status should be 200').equal(200);
+      };
+
+      await retryRequest(req, assertions);
+    });
   });
 
   regexWrongPatterns.forEach((wrongPattern) => {
@@ -263,26 +299,15 @@ describe('@smoke: Router Functionality Tests', function () {
     });
   });
 
-  regexCorrectPatterns.forEach((correctPattern) => {
-    it(`should route a request with matching regex path: ${correctPattern}`, async function () {
-      const req = () =>
-        axios({
-          url: `${proxyUrl}${correctPattern}`,
-        });
-
-      const assertions = (resp) => {
-        expect(resp.status, 'Status should be 200').equal(200);
-      };
-
-      await retryRequest(req, assertions);
-    });
-  });
-
   it('should PATCH the route host', async function () {
+    // *** KOKO DOES NOT PATCH REQUESTS ***
     const resp = await axios({
-      method: 'patch',
+      method: isGateway() ? 'patch' : 'put',
       url: `${routesUrl}/${routeId}`,
       data: {
+        service: {
+          id: serviceDetails.id
+        },
         hosts: ['test'],
         paths: [routePayload.paths[0]],
       },
@@ -310,11 +335,15 @@ describe('@smoke: Router Functionality Tests', function () {
   });
 
   it('should route the request with correct host', async function () {
-    let resp = await axios({
-      method: 'patch',
+    const resp = await axios({
+      method: isGateway() ? 'patch' : 'put',
       url: `${routesUrl}/${routeId}`,
       data: {
-        hosts: [currentHost],
+        service: {
+          id: serviceDetails.id
+        },
+        paths: [routePayload.paths[0]],
+        hosts: [currentHost]
       },
       headers: { 'Content-Type': 'application/json' },
     });
@@ -327,52 +356,58 @@ describe('@smoke: Router Functionality Tests', function () {
 
     await waitForConfigRebuild();
 
-    resp = await axios(`${proxyUrl}${routePayload.paths[0]}`);
-    expect(resp.status, 'Status should be 200').equal(200);
-    logResponse(resp);
+    const req = () => axios(`${proxyUrl}${routePayload.paths[0]}`);
+    const assertions = (resp) => {
+      expect(resp.status, 'Status should be 200').equal(200);
+    };
+
+    await retryRequest(req, assertions);
   });
 
-  it('should delete the route by name', async function () {
-    const resp = await axios({
-      method: 'delete',
-      url: `${routesUrl}/${routePayload.name}`,
+  if(isGateway()) {
+    it('should delete the route by name', async function () {
+      const resp = await axios({
+        method: 'delete',
+        url: `${routesUrl}/${routePayload.name}`,
+      });
+  
+      logResponse(resp);
+      expect(resp.status, 'Status should be 204').to.equal(204);
     });
 
-    logResponse(resp);
-    expect(resp.status, 'Status should be 204').to.equal(204);
-  });
-
-  it('should not panic and create a route with long regex path', async function () {
-    // generate string longer than 2048 bytes
-    const path = 'x'.repeat(3 * 1024);
-
-    const resp = await postNegative(
-      `${adminUrl}/services/${serviceDetails.id}/routes`,
-      {
-        name: randomString(),
-        paths: [`~/${path}/[^\\/]{14}()$`],
-      },
-      'post',
-      {},
-      { rejectUnauthorized: true }
-    );
-    logResponse(resp);
-
-    routeId = resp.data.id;
-
-    expect(resp.status, 'Status should be 201').to.equal(201);
-
-    await wait(4000); // eslint-disable-line no-restricted-syntax
-    const currentLogs = getGatewayContainerLogs(kongContainerName, 15);
-    const panickLog = findRegex('panicked', currentLogs);
-    const outOfRangeLog = findRegex('out of range', currentLogs);
-
-    expect(panickLog, 'Should not see router panic error log').to.be.false;
-    expect(outOfRangeLog, 'Should not see out of range error log').to.be.false;
-  });
+    it('should not panic and create a route with long regex path', async function () {
+      // generate string longer than 2048 bytes
+      // *** KOKO MAX BYTES IS 1024 AND IT DOESN'T RECOGNIZE "\\/" ESCAPE SEQUENCE ***
+      const path = 'x'.repeat(3072);
+  
+      const resp = await postNegative(
+        `${adminUrl}/services/${serviceDetails.id}/routes`,
+        {
+          name: randomString(),
+          paths: [`~/${path}/[^\\/]{14}()$`],
+        },
+        'post',
+        {},
+        { rejectUnauthorized: true }
+      );
+      logResponse(resp);
+  
+      routeId = resp.data.id;
+  
+      expect(resp.status, 'Status should be 201').to.equal(201);
+  
+      await wait(4000); // eslint-disable-line no-restricted-syntax
+      const currentLogs = getGatewayContainerLogs(kongContainerName, 15);
+      const panickLog = findRegex('panicked', currentLogs);
+      const outOfRangeLog = findRegex('out of range', currentLogs);
+  
+      expect(panickLog, 'Should not see router panic error log').to.be.false;
+      expect(outOfRangeLog, 'Should not see out of range error log').to.be.false;
+    });
+  }
 
   after(async function () {
     await deleteGatewayRoute(routeId);
-    await deleteGatewayService(serviceName);
+    await deleteGatewayService(isGateway() ? serviceName : serviceDetails.id);
   });
 });
